Handle driver api errors (#10)
* Don't raise exceptions from Contrail Driver methods

Don't raise exception/errors in the Contrail Driver implementation
methods for LBaaSv2 APIs. Raising an exception/error is causing
contrail-svc-monitor process to crash.

* Handle not-found error in delete LB

In Contrail Driver, when delete loadbalancer is invoked, we are checking
the LB object in Cotrail Config DB. Because of asynchronous method of
API handling in Contrail, object from config DB is sometimes deleted
even before Driver can handle the method. The Driver will then fail to
fetch the object from DB and get into null-pointer kind of error.

Handle this error by relying on the data passed to the driver if the
driver cannot find the object in DB. The data passed to driver has a
copy of LB details that can be used to delete the objects from Avi.

* Log LB object not found

+++ avi_lbaasv2/common/avi_generic.py
@@ -358,10 +358,21 @@ def update_vsvip(os_lb, avi_client, avi_tenant_uuid, cloud, vsvip=None,
     return res
 
 
-def delete_vsvip(os_lb, avi_client):
-    vsvip_uuid = form_vsvip_uuid(os_lb.id)
-    avi_tenant_uuid = os2avi_uuid("tenant", os_lb.tenant_id)
-    avi_client.delete("vsvip", vsvip_uuid, avi_tenant_uuid)
+def delete_vsvip(os_lb, avi_client, contrail_lb=None):
+    vsvip_uuid = None
+    avi_tenant_uuid = None
+    if os_lb:
+        vsvip_uuid = form_vsvip_uuid(os_lb.id)
+        avi_tenant_uuid = os2avi_uuid("tenant", os_lb.tenant_id)
+    elif contrail_lb:
+        lb_id = contrail_lb.get('id', None)
+        tenant_id = contrail_lb.get('tenant_id', None)
+        if lb_id and tenant_id:
+            vsvip_uuid = form_vsvip_uuid(lb_id)
+            avi_tenant_uuid = os2avi_uuid("tenant", tenant_id)
+
+    if vsvip_uuid and avi_tenant_uuid:
+        avi_client.delete("vsvip", vsvip_uuid, avi_tenant_uuid)
 
 
 def get_vrf_context(subnet_uuid, cloud, avi_tenant_uuid, avi_client,
